refactor : common NodeValue between direct and bytecode interpreters

+++ test/src/300Interpreter_test.cpp
@@ -11,7 +11,7 @@
 #include "Interpreter/Service/ScriptBinding.h"
 #include "Interpreter/Value/Script.h"
 #include "Interpreter/ScriptCache.h"
-#include "NodeValue/ObjectMemory.h"
+#include "Runtime/Value/ObjectMemory.h"
 #include "Service/TypeCrosser/TypeCrossExpression.h"
 
 const auto reservedKeywords = ska::ReservedKeywordsPool{};
@@ -48,30 +48,30 @@ TEST_CASE("[Interpreter]") {
 		SUBCASE("Basic Maths") {
 			auto astPtr = ASTFromInputSemanticTCInterpreter("(4 * 5) + 2 * (3 + 4 - 1) + 1 + 9;", data);
 			auto res = data.interpreter->script(astPtr);
-            CHECK(res.nodeval<int>() == 42);
+            CHECK(res.nodeval<long>() == 42);
 		}
 		SUBCASE("Basic Maths with var") {
 			auto astPtr = ASTFromInputSemanticTCInterpreter("var toto = 4; toto; (toto * 5) + 2 * (3 + 4 - 1) + 1 + 9;", data);
 			auto res = data.interpreter->script(astPtr);
-			CHECK(res.nodeval<int>() == 42);
+			CHECK(res.nodeval<long>() == 42);
 		}
 
 		SUBCASE("Var declared") {
 			auto astPtr = ASTFromInputSemanticTCInterpreter("var toto = 14; toto;", data);
 			auto res = data.interpreter->script(astPtr);
-			CHECK(res.nodeval<int>() == 14);
+			CHECK(res.nodeval<long>() == 14);
 		}
 
 		SUBCASE("Var reaffected") {
 			auto astPtr = ASTFromInputSemanticTCInterpreter("var toto = 4; toto; toto = 25;", data);
 			auto res = data.interpreter->script(astPtr);
-			CHECK(res.nodeval<int>() == 25);
+			CHECK(res.nodeval<long>() == 25);
 		}
 
 		SUBCASE("Var reaffected using the same var (int)") {
 			auto astPtr = ASTFromInputSemanticTCInterpreter("var toto = 4; toto; toto = toto * 2;", data);
 			auto res = data.interpreter->script(astPtr);
-			CHECK(res.nodeval<int>() == 8);
+			CHECK(res.nodeval<long>() == 8);
 		}
 
 		SUBCASE("Var reaffected using the same var (string)") {
@@ -101,37 +101,37 @@ TEST_CASE("[Interpreter]") {
 		SUBCASE("For loop statement") {
 			auto astPtr = ASTFromInputSemanticTCInterpreter("var tititi = 2; for(var i = 0; i < 5; i = i + 1) { tititi = tititi + i; } tititi; ", data);
 			auto res = data.interpreter->script(astPtr);
-			CHECK(res.nodeval<int>() == 12);
+			CHECK(res.nodeval<long>() == 12);
 		}
 
         SUBCASE("If block true") {
             auto astPtr = ASTFromInputSemanticTCInterpreter("var toto = 1; if(toto + 1 == 2) { toto = 123; } toto;", data);
 			auto res = data.interpreter->script(astPtr);
-			CHECK(res.nodeval<int>() == 123);
+			CHECK(res.nodeval<long>() == 123);
         }
 
 		SUBCASE("If block false") {
 			auto astPtr = ASTFromInputSemanticTCInterpreter("var toto = 1; if(toto + 1 == 12) { toto = 123; } toto;", data);
 			auto res = data.interpreter->script(astPtr);
-			CHECK(res.nodeval<int>() == 1);
+			CHECK(res.nodeval<long>() == 1);
 		}
 
 		SUBCASE("If block pure true") {
 			auto astPtr = ASTFromInputSemanticTCInterpreter("var toto = 0; if(true) { toto = 123; } toto;", data);
 			auto res = data.interpreter->script(astPtr);
-			CHECK(res.nodeval<int>() == 123);
+			CHECK(res.nodeval<long>() == 123);
 		}
 
 		SUBCASE("If Else block if false") {
 			auto astPtr = ASTFromInputSemanticTCInterpreter("var toto = 1; if(toto + 1 == 12) { toto = 123; } else { toto = 24; } toto;", data);
 			auto res = data.interpreter->script(astPtr);
-			CHECK(res.nodeval<int>() == 24);
+			CHECK(res.nodeval<long>() == 24);
 		}
 
 		SUBCASE("If Else block if true") {
 			auto astPtr = ASTFromInputSemanticTCInterpreter("var toto = 1; if(toto + 1 == 2) { toto = 123; } else { toto = 24; } toto;", data);
 			auto res = data.interpreter->script(astPtr);
-			CHECK(res.nodeval<int>() == 123);
+			CHECK(res.nodeval<long>() == 123);
 		}
 
 		SUBCASE("Var reaffected string with number") {
@@ -143,86 +143,86 @@ TEST_CASE("[Interpreter]") {
 		SUBCASE("Array of int") {
 			auto astPtr = ASTFromInputSemanticTCInterpreter("var toto = [14, 25, 13, 2]; toto[1];", data);
 			auto res = data.interpreter->script(astPtr);
-			CHECK(res.nodeval<int>() == 25);
+			CHECK(res.nodeval<long>() == 25);
 		}
 
 		SUBCASE("Array of int : add a cell") {
 			auto astPtr = ASTFromInputSemanticTCInterpreter("var toto = [14, 25]; toto = toto + 4;", data);
 			auto res = data.interpreter->script(astPtr).nodeval<ska::NodeValueArray>();
 			CHECK(res->size() == 3);
-			CHECK((*res)[0] == 14);
-			CHECK((*res)[1] == 25);
-			CHECK((*res)[2] == 4);
+			CHECK((*res)[0] == 14L);
+			CHECK((*res)[1] == 25L);
+			CHECK((*res)[2] == 4L);
 		}
 
 		SUBCASE("Array of int : del a cell") {
 			auto astPtr = ASTFromInputSemanticTCInterpreter("var toto = [14, 25, 13, 2]; toto = toto - 2;", data);
 			auto res = data.interpreter->script(astPtr).nodeval<ska::NodeValueArray>();
 			CHECK(res->size() == 2);
-			CHECK((*res)[0] == 14);
-			CHECK((*res)[1] == 25);
+			CHECK((*res)[0] == 14L);
+			CHECK((*res)[1] == 25L);
 		}
 
 		SUBCASE("Array of int : del a range of cells") {
 			auto astPtr = ASTFromInputSemanticTCInterpreter("var toto = [14, 25, 13, 2]; toto = toto - [1, 2];", data);
 			auto res = data.interpreter->script(astPtr).nodeval<ska::NodeValueArray>();
 			CHECK(res->size() == 2);
-			CHECK((*res)[0] == 14);
-			CHECK((*res)[1] == 2);
+			CHECK((*res)[0] == 14L);
+			CHECK((*res)[1] == 2L);
 		}
 
 		SUBCASE("Assigning a cell of an array of int") {
 			auto astPtr = ASTFromInputSemanticTCInterpreter("var toto = [14, 25, 13, 2]; toto[1] = 226; toto[1];", data);
 			auto res = data.interpreter->script(astPtr);
-			CHECK(res.nodeval<int>() == 226);
+			CHECK(res.nodeval<long>() == 226);
 		}
 
 		SUBCASE("Basic function") {
 			auto astPtr = ASTFromInputSemanticTCInterpreter("var totoFunc = function() : int { return 3; }; var titi = totoFunc(); titi;", data);
 			auto res = data.interpreter->script(astPtr);
-			CHECK(res.nodeval<int>() == 3);
+			CHECK(res.nodeval<long>() == 3);
 		}
 		
 		SUBCASE("Function 1 parameter") {
 			auto astPtr = ASTFromInputSemanticTCInterpreter("var totoFunc = function(value: int) : int { return value + 3; }; var titi = totoFunc(7); titi;", data);
 			auto res = data.interpreter->script(astPtr);
-			CHECK(res.nodeval<int>() == 10);
+			CHECK(res.nodeval<long>() == 10);
 		}
 
 		SUBCASE("Function 1 parameter 2x calls") {
 			auto astPtr = ASTFromInputSemanticTCInterpreter("var totoFunc = function(value: int) : int { return value + 3; }; var titi = totoFunc(7); titi = totoFunc(17);", data);
 			auto res = data.interpreter->script(astPtr);
-			CHECK(res.nodeval<int>() == 20);
+			CHECK(res.nodeval<long>() == 20);
 		}
 
 		SUBCASE("Function 1 parameter using a parent scope var") {
 			auto astPtr = ASTFromInputSemanticTCInterpreter("var test = 20; var totoFunc = function(value: int) : int { return test + value + 3; }; var titi = totoFunc(7); titi;", data);
 			auto res = data.interpreter->script(astPtr);
-			CHECK(res.nodeval<int>() == 30);
+			CHECK(res.nodeval<long>() == 30);
 		}
 
 		SUBCASE("Function 0 parameter creating custom object") {
 			auto astPtr = ASTFromInputSemanticTCInterpreter("var totoFunc = function() : var { return { num: 3, test: \"test\" }; }; var titi = totoFunc(); titi.num;", data);
 			auto res = data.interpreter->script(astPtr);
-			CHECK(res.nodeval<int>() == 3);
+			CHECK(res.nodeval<long>() == 3);
 		}
 			
 		SUBCASE("Function 0 parameter creating custom object with use as parameter of another one") {
 			auto astPtr = ASTFromInputSemanticTCInterpreter("var Dummy = function(): var { return { data: 3 }; }; var Runner = function(): var { var print = function(i: Dummy()): int { return i.data; }; return { print : print }; }; Runner().print(Dummy());", data);
 			auto res = data.interpreter->script(astPtr);
-			CHECK(res.nodeval<int>() == 3);
+			CHECK(res.nodeval<long>() == 3);
 		}
 
 		SUBCASE("Array use from field access") {
 			auto astPtr = ASTFromInputSemanticTCInterpreter("var Coldragon = function() : var{ var array = [20, 150, 2]; return { array : array }; }; var target = Coldragon(); var data = target.array[0] - 88; data;", data);
 			auto res = data.interpreter->script(astPtr);
-			CHECK(res.nodeval<int>() == -68);
+			CHECK(res.nodeval<long>() == -68);
 		}
 
 		SUBCASE("double array string : cell") {
 			auto astPtr = ASTFromInputSemanticTCInterpreter("var str167 = [[0, 1], [2, 3]]; str167[0]; str167[0][1];", data);
 			auto res = data.interpreter->script(astPtr);
-			CHECK(res.nodeval<int>() == 1);
+			CHECK(res.nodeval<long>() == 1);
 		}
 
 		SUBCASE("using a function as a parameter") {
@@ -236,7 +236,7 @@ TEST_CASE("[Interpreter]") {
 				"var object = lvalFunc209(lvalFunc206);"
 				"object.test;", data);
 			auto res = data.interpreter->script(astPtr);
-			CHECK(res.nodeval<int>() == 14);
+			CHECK(res.nodeval<long>() == 14);
 		}
 
 		SUBCASE("using a callback function as a parameter without using the source type (function type compatibility)") {
@@ -249,7 +249,7 @@ TEST_CASE("[Interpreter]") {
 				"var callback = function() { testOut = 789; };"
 				"lvalFunc219(callback);", data);
 			auto res = data.interpreter->script(astPtr);
-			CHECK(res.nodeval<int>() == 789);
+			CHECK(res.nodeval<long>() == 789);
 		}
 
 		SUBCASE("for with empty statement in if") {
